+++ ui/src/components/navigation/mobile/MobileNav.tsx
@@ -1,13 +1,18 @@
 import { FC, Fragment, useState } from "react";
 import { Button, Drawer, Menu } from "react-daisyui";
-import { Menu as MenuIcon } from "lucide-react";
+import { Menu as MenuIcon, ChevronDown } from "lucide-react";
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { useAuth0 } from "@auth0/auth0-react";
+// ..custom
 import { INavItem } from "@/components/navigation/navigation.list";
+import UserProfileDropdownWrapper from "../desktop/UserProfileDropdown";
 
 type IMobileNavbarProps = {
   navItems: INavItem[];
 };
 
 const MobileNavbar: FC<IMobileNavbarProps> = () => {
+  const { isAuthenticated, loginWithRedirect, user, logout } = useAuth0();
   const [drawerOpened, setDrawerOpened] = useState(false);
 
   return (
@@ -17,11 +22,11 @@ const MobileNavbar: FC<IMobileNavbarProps> = () => {
           open={drawerOpened}
           onClickOverlay={() => setDrawerOpened(!drawerOpened)}
           side={
-            <Menu className="min-h-full w-80 gap-2 bg-base-100 p-4 text-base-content">
+            <Menu className="min-h-full gap-2 p-4 w-80 bg-base-100 text-base-content">
               <Menu.Item className="font-medium">
                 <a
-                  href="index.html"
-                  className="text-brand-gradient text-xl font-bold tracking-tighter"
+                  href="#"
+                  className="text-xl font-bold tracking-tighter text-brand-gradient"
                 >
                   MANNA Group International
                 </a>
@@ -31,37 +36,61 @@ const MobileNavbar: FC<IMobileNavbarProps> = () => {
                 className="font-medium"
                 onClick={() => setDrawerOpened(false)}
               >
-                <a href="#/">Home</a>
+                <a href="#/about/who-we-are">About</a>
               </Menu.Item>
               <Menu.Item
                 className="font-medium"
                 onClick={() => setDrawerOpened(false)}
               >
-                <a href="#/about/10">About</a>
+                <a href="#/products/shop-products">SHOP Products!</a>
               </Menu.Item>
               <Menu.Item
                 className="font-medium"
                 onClick={() => setDrawerOpened(false)}
               >
-                <a href="#/products">Products</a>
+                <a href="#/company/gallery-section">Company</a>
               </Menu.Item>
               <Menu.Item
                 className="font-medium"
                 onClick={() => setDrawerOpened(false)}
               >
-                <a href="#/company/10">Company</a>
+                <a href="#/contact/contact-us-section">Contact Us</a>
               </Menu.Item>
-              <Menu.Item
-                className="font-medium"
-                onClick={() => setDrawerOpened(false)}
-              >
-                <a href="#/contact">Contact Us</a>
+              {/* UserProfile */}
+              <Menu.Item className="font-medium dropdown">
+                {isAuthenticated ? (
+                  <UserProfileDropdownWrapper>
+                    <Fragment>
+                      <Avatar>
+                        <AvatarImage src={user?.picture} />
+                        <AvatarFallback>
+                          {user?.name?.substring(0, 1).toUpperCase()}
+                        </AvatarFallback>
+                      </Avatar>
+                      <ChevronDown />
+                    </Fragment>
+                  </UserProfileDropdownWrapper>
+                ) : (
+                  <Button
+                    onClick={() => loginWithRedirect()}
+                    size={"sm"}
+                    color={"primary"}
+                  >
+                    Login
+                  </Button>
+                )}
               </Menu.Item>
-              <a href="#download" onClick={() => setDrawerOpened(false)}>
-                <Button size={"sm"} color={"primary"}>
-                  Download Now
+              {/* Logout */}
+              { isAuthenticated && <Menu.Item className="font-medium dropdown">
+                <Button
+                  onClick={() => logout()}
+                  size={"sm"}
+                  color={"primary"}
+                >
+                  Logout
                 </Button>
-              </a>
+              </Menu.Item>
+              }
             </Menu>
           }
         >