+++ ui/src/pages/company/sections/Gallery.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 const GallerySection: FC = () => {
   return (
-    <section className="py-8 lg:py-24" id="gallery-events">
+    <section className="py-8 lg:py-24" id="gallery-section">
       <div className="container">
         {/* get together */}
         <div className="grid gap-16 lg:grid-cols-2 xl:gap-24">
@@ -10,10 +10,10 @@ const GallerySection: FC = () => {
             <div className="flex justify-center">{/* Gallery Section */}</div>
           </div>
           <div className="col-span">
-            <h1 className="text-center text-3xl/tight font-bold leading-10 tracking-tight sm:text-start lg:text-4xl/tight">
+            <h1 className="font-bold leading-10 tracking-tight text-center text-3xl/tight sm:text-start lg:text-4xl/tight">
               Gallery
             </h1>
-            <p className="mt-3 font-semibold text-brand-gradient text-xl">
+            <p className="mt-3 text-xl font-semibold text-brand-gradient">
               MANNA Get together 2022
             </p>
             <p className="mt-3 text-base">
@@ -22,12 +22,12 @@ const GallerySection: FC = () => {
           </div>
         </div>
         {/* mayor visit */}
-        <div className="mt-8 grid gap-16 lg:grid-cols-2 xl:gap-24">
+        <div className="grid gap-16 mt-8 lg:grid-cols-2 xl:gap-24">
           <div className="relative">
             <div className="flex justify-center">{/* Gallery Section */}</div>
           </div>
           <div className="col-span">
-            <p className="mt-3 font-semibold text-xl text-brand-gradient">
+            <p className="mt-3 text-xl font-semibold text-brand-gradient">
               Press conference with Mayor of Trenton, NJ
             </p>
             <p className="mt-3 text-base">
@@ -37,12 +37,12 @@ const GallerySection: FC = () => {
           </div>
         </div>
         {/* hotel ramada */}
-        <div className="mt-8 grid gap-16 lg:grid-cols-2 xl:gap-24">
+        <div className="grid gap-16 mt-8 lg:grid-cols-2 xl:gap-24">
           <div className="relative">
             <div className="flex justify-center">{/* Gallery Section */}</div>
           </div>
           <div className="col-span">
-            <p className="mt-3 font-semibold text-xl text-brand-gradient">
+            <p className="mt-3 text-xl font-semibold text-brand-gradient">
               Launch Meeting at Hotel Ramada, East Orange, NJ
             </p>
             <p className="mt-3 text-base">
@@ -52,12 +52,12 @@ const GallerySection: FC = () => {
           </div>
         </div>
         {/* hotel rccg */}
-        <div className="mt-8 grid gap-16 lg:grid-cols-2 xl:gap-24">
+        <div className="grid gap-16 mt-8 lg:grid-cols-2 xl:gap-24">
           <div className="relative">
             <div className="flex justify-center">{/* Gallery Section */}</div>
           </div>
           <div className="col-span">
-            <p className="mt-3 font-semibold text-xl text-brand-gradient">
+            <p className="mt-3 text-xl font-semibold text-brand-gradient">
               Launch Meeting at RCCG Dominion Cathedral Orange, NJ
             </p>
             <p className="mt-3 text-base">
@@ -67,12 +67,12 @@ const GallerySection: FC = () => {
           </div>
         </div>
         {/* meeting with investors */}
-        <div className="mt-8 grid gap-16 lg:grid-cols-2 xl:gap-24">
+        <div className="grid gap-16 mt-8 lg:grid-cols-2 xl:gap-24">
           <div className="relative">
             <div className="flex justify-center">{/* Gallery Section */}</div>
           </div>
           <div className="col-span">
-            <p className="mt-3 font-semibold text-xl text-brand-gradient">
+            <p className="mt-3 text-xl font-semibold text-brand-gradient">
               First Launch Meeting with Investors at Morris Avenue, Union
               office, NJ
             </p>