+++ ui/src/pages/company/sections/International.tsx
@@ -4,18 +4,18 @@ import WorldMapChart from "@/components/maps/WorldMapChart";
 
 export const InternationalSection: FC = () => {
   return (
-    <section className="relative py-8 lg:py-24" id="home">
-      <div className="absolute inset-0 bg-cover bg-center bg-no-repeat dark:opacity-5"></div>
+    <section className="relative py-8 lg:py-24" id="usa-international">
+      <div className="absolute inset-0 bg-center bg-no-repeat bg-cover dark:opacity-5"></div>
       <div className="container relative z-10">
-        <div className="mt-16 grid items-center gap-12 lg:grid-cols-2 xl:gap-36">
+        <div className="grid items-center gap-12 mt-16 lg:grid-cols-2 xl:gap-36">
           <div className="order-2 lg:order-1">
             <WorldMapChart />
           </div>
           <div className="order-1 lg:order-2">
-            <h1 className="text-center text-3xl/tight font-bold leading-10 tracking-tight sm:text-start lg:text-4xl/tight">
+            <h1 className="font-bold leading-10 tracking-tight text-center text-3xl/tight sm:text-start lg:text-4xl/tight">
               <span className="text-brand-gradient">MANNA</span> International
             </h1>
-            <p className="mt-8 text-center text-base sm:text-start">
+            <p className="mt-8 text-base text-center sm:text-start">
               We at MANNA are planning to globalize the products developed by us
               for the benefit of our customers and patients world over. We aim
               to be a globally recognized entity, through strategic alliances in