+++ ui/src/pages/contact/Contact.page.tsx
@@ -1,11 +1,25 @@
-import { FC } from "react";
-import mobile3Img from "@/assets/images/landing/mobile-3.png";
+import { FC, useEffect } from "react";
+import { useParams } from "react-router-dom";
 import { Globe, Factory, PhoneIcon, MonitorSmartphoneIcon } from "lucide-react";
 import { Card } from "react-daisyui";
+// ..custom
+import mobile3Img from "@/assets/images/landing/mobile-3.png";
 
 const ContactPage: FC = () => {
+  const { id } = useParams();
+  
+  useEffect(()=>{
+    if (Boolean(id) && id !== "") {      
+      document.getElementById(id)?.scrollIntoView({
+        behavior: 'smooth',
+        block: 'start',
+        inline: 'nearest'
+      });    
+    }
+  },[id]);
+
   return (
-    <section className="py-8 lg:py-24" id="organize">
+    <section className="py-8 lg:py-24" id="contact-us-section">
       <div className="container">
         <div className="grid gap-16 lg:grid-cols-2 xl:gap-24">
           <div className="order-2 lg:order-1">
@@ -15,7 +29,7 @@ const ContactPage: FC = () => {
             </p>
             <div className="mt-8 space-y-8">
               <div className="flex items-start gap-5">
-                <div className="inline-flex p-2 items-center justify-center rounded bg-primary/20 font-medium text-primary">
+                <div className="inline-flex items-center justify-center p-2 font-medium rounded bg-primary/20 text-primary">
                   <Globe size={24} />
                 </div>
                 <div>
@@ -26,7 +40,7 @@ const ContactPage: FC = () => {
                 </div>
               </div>
               <div className="flex items-start gap-5">
-                <div className="inline-flex p-2 items-center justify-center rounded bg-primary/20 font-medium text-primary">
+                <div className="inline-flex items-center justify-center p-2 font-medium rounded bg-primary/20 text-primary">
                   <Factory size={24} />
                 </div>
                 <div>
@@ -36,8 +50,8 @@ const ContactPage: FC = () => {
                   </p>
                 </div>
               </div>
-              <div className="flex items-start gap-5">
-                <div className="inline-flex p-2 items-center justify-center rounded bg-primary/20 font-medium text-primary">
+              <div className="flex items-start gap-5" id="send-inquiry">
+                <div className="inline-flex items-center justify-center p-2 font-medium rounded bg-primary/20 text-primary">
                   <PhoneIcon size={24} />
                 </div>
                 <div>
@@ -58,13 +72,13 @@ const ContactPage: FC = () => {
               <img alt="Mobile 3" className="h-[500px]" src={mobile3Img} />
             </div>
             <div className="absolute end-10 top-[30%]">
-              <Card className="bg-base-100 shadow transition-all hover:shadow-lg">
+              <Card className="transition-all shadow bg-base-100 hover:shadow-lg">
                 <Card.Body className="flex flex-row items-center gap-3 p-3">
-                  <div className="flex items-center justify-center rounded-full bg-primary p-2">
+                  <div className="flex items-center justify-center p-2 rounded-full bg-primary">
                     <MonitorSmartphoneIcon className="size-6 text-primary-content" />
                   </div>
                   <div>
-                    <p className="text-xl/none font-semibold">Contact us</p>
+                    <p className="font-semibold text-xl/none">Contact us</p>
                     <p className="">We're here to help !!</p>
                   </div>
                 </Card.Body>