+++ ui/src/pages/products/Products.page.tsx
@@ -1,26 +1,35 @@
 import { FC, useEffect } from "react";
+import { useParams } from "react-router-dom";
 import { Button, Card } from "react-daisyui";
 import { ShoppingCart, ArrowRightIcon } from "lucide-react";
 // ..images
 import { productList, IProduct } from "./data/product_list";
-import { scrollToOffset } from "@/util/helper";
 
 const ProductsPage: FC = () => {
-  useEffect(() => {
-    scrollToOffset(Number(2));
-  }, []);
+
+  const { id } = useParams();
+  
+  useEffect(()=>{
+    if (Boolean(id) && id !== "") {      
+      document.getElementById(id)?.scrollIntoView({
+        behavior: 'smooth',
+        block: 'start',
+        inline: 'nearest'
+      });    
+    }
+  },[id]);
 
   return (
-    <section className="relative py-8 lg:py-24" id="products_container">
+    <section className="relative py-8 lg:py-24" id="shop-products">
       <div className="container relative z-10">
         {/* Products section*/}
         {productList.map((item: IProduct, idx: number) => (
           <Card
             key={idx + 1}
-            className="border border-base-content/10 rounded-xl p-4 my-4"
+            className="p-4 my-4 border border-base-content/10 rounded-xl"
           >
             <div className="grid items-center gap-12 lg:grid-cols-2 xl:gap-36">
-              <div className="order-1 flex justify-center lg:justify-start">
+              <div className="flex justify-center order-1 lg:justify-start">
                 <img
                   alt={item.title.toLowerCase()}
                   className="h-[500px] rounded-xl"
@@ -28,23 +37,23 @@ const ProductsPage: FC = () => {
                 />
               </div>
               <div className="order-2">
-                <h1 className="text-center text-3xl/tight font-bold leading-10 tracking-tight sm:text-start lg:text-4xl/tight">
+                <h1 className="font-bold leading-10 tracking-tight text-center text-3xl/tight sm:text-start lg:text-4xl/tight">
                   <span className="text-brand-gradient">{item.name}</span>
                 </h1>
-                <p className="mt-4 text-center text-base sm:text-start font-semibold">
+                <p className="mt-4 text-base font-semibold text-center sm:text-start">
                   {item.title}
                 </p>
-                <p className="mt-4 text-center text-base sm:text-start">
+                <p className="mt-4 text-base text-center sm:text-start">
                   {item.description}
                 </p>
-                <p className="mt-8 font-semibold text-lg">
+                <p className="mt-8 text-lg font-semibold">
                   Price:{" "}
                   <span className="text-2xl text-brand-gradient">
                     ${item.price}
                   </span>
                 </p>
                 <section className="flex justify-center lg:justify-end">
-                  <Button color={"primary"} size={"sm"} className="mt-8 mx-2">
+                  <Button color={"primary"} size={"sm"} className="mx-2 mt-8">
                     <ShoppingCart size={16} />
                     Add to Cart
                   </Button>
@@ -59,7 +68,7 @@ const ProductsPage: FC = () => {
         ))}
 
         {/* Promo Video section*/}
-        <div className="mt-16 grid items-center gap-12 lg:grid-cols-2 xl:gap-36">
+        <div className="grid items-center gap-12 mt-16 lg:grid-cols-2 xl:gap-36">
           <div className="order-2 lg:order-1">
             <iframe
               width="560"
@@ -73,10 +82,10 @@ const ProductsPage: FC = () => {
             ></iframe>
           </div>
           <div className="order-1 lg:order-2">
-            <h1 className="text-center text-3xl/tight font-bold leading-10 tracking-tight sm:text-start lg:text-4xl/tight">
+            <h1 className="font-bold leading-10 tracking-tight text-center text-3xl/tight sm:text-start lg:text-4xl/tight">
               <span className="text-brand-gradient">MANNA</span> Our Products
             </h1>
-            <p className="mt-8 text-center text-base sm:text-start">
+            <p className="mt-8 text-base text-center sm:text-start">
               Watch our promotional video for the overview of all products
             </p>
           </div>