+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v0.5.1
+- 10/06/2024
+- Fix all navigation issues
+- Fix typescript generating random .js files
+- Fix Mobile sidebar issue
+- Fix Footer navigation issues
+- Refactor `ClientRouter.tsx`
+
 ## v0.5.0
 - 10/05/2024
 - Deployed API to Fly.io